+++ tests/posix/clock_selection/prj.conf
@@ -1,0 +1,9 @@
+CONFIG_POSIX_API=y
+CONFIG_ZTEST=y
+
+CONFIG_POSIX_AEP_CHOICE_BASE=y
+CONFIG_POSIX_CLOCK_SELECTION=y
+
+CONFIG_DYNAMIC_THREAD=y
+CONFIG_DYNAMIC_THREAD_POOL_SIZE=3
+CONFIG_THREAD_STACK_INFO=y
