+++ tests/posix/clock_selection/testcase.yaml
@@ -1,0 +1,28 @@
+common:
+  filter: not CONFIG_NATIVE_LIBC
+  tags:
+    - posix
+    - clock_selection
+  # 1 tier0 platform per supported architecture
+  platform_key:
+    - arch
+    - simulation
+  integration_platforms:
+    - qemu_riscv64
+  min_flash: 64
+  min_ram: 32
+tests:
+  portability.posix.clock_selection: {}
+  portability.posix.clock_selection.minimal:
+    extra_configs:
+      - CONFIG_MINIMAL_LIBC=y
+  portability.posix.clock_selection.newlib:
+    filter: TOOLCHAIN_HAS_NEWLIB == 1
+    extra_configs:
+      - CONFIG_NEWLIB_LIBC=y
+      - CONFIG_NEWLIB_LIBC_MIN_REQUIRED_HEAP_SIZE=8192
+  portability.posix.clock_selection.picolibc:
+    tags: picolibc
+    filter: CONFIG_PICOLIBC_SUPPORTED
+    extra_configs:
+      - CONFIG_PICOLIBC=y
