+++ tests/posix/common/src/cond.c
@@ -49,22 +49,9 @@ ZTEST(cond, test_cond_resource_leak)
 
 ZTEST(cond, test_pthread_condattr)
 {
-	clockid_t clock_id;
 	pthread_condattr_t att = {0};
 
 	zassert_ok(pthread_condattr_init(&att));
-
-	zassert_ok(pthread_condattr_getclock(&att, &clock_id), "pthread_condattr_getclock failed");
-	zassert_equal(clock_id, CLOCK_REALTIME, "clock attribute not set correctly");
-
-	zassert_ok(pthread_condattr_setclock(&att, CLOCK_REALTIME),
-		   "pthread_condattr_setclock failed");
-
-	zassert_ok(pthread_condattr_getclock(&att, &clock_id), "pthread_condattr_setclock failed");
-	zassert_equal(clock_id, CLOCK_REALTIME, "clock attribute not set correctly");
-
-	zassert_equal(pthread_condattr_setclock(&att, 42), -EINVAL,
-		      "pthread_condattr_setclock did not return EINVAL");
 
 	zassert_ok(pthread_condattr_destroy(&att));
 }
